% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libload.R
\name{libload}
\alias{libload}
\title{Load all libraries from an R script and install them as needed}
\usage{
libload(file, ...)
}
\arguments{
\item{file}{R script file to load}

\item{...}{library_pattern to be passed into .parse_r_script}
}
\description{
This function will take a R script, load any packages, and install packages
as needed
}
\examples{
\dontrun{
libload('r_script_file.R')
}
}

